Added minor css changes and fixed pr reqs
